Ongoing work to cleanup the 'present' tool HTML
I'd like sublevels to work.  Issues filed on this say "it won't happen"
or "do it yourself".  Sigh.

I've created 'test.slide' for use in testing changes, and
'present.patch' for corrections already made and working (such as code
being shown in blue, and font styling working in titles).  The README.md
describes how to apply the (optional) patch to the 'present' tool.