Updated test slidedeck for new format
We now support using '^* ' to indicate a new chapter and the chapter
number will be printed immediately in front of the (styled) title
string.  Then, '^** ' can be used to indicate the start of a new slide
within the chapter.  It will be prefixed with a two-level outline-style
section number, such as 1.1, 1.2, 2.5, or 10.3.  (The slide titles are
also styled.)  When run through the template system, all top-level
chapter titles produce a new slide and all subelements start new slides
(so it's best not to put anything under the top-level chapter titles
except sublevel slide titles).  Such as:

    * Chapter One

    ** Slide 1

    Text here
    - And bullets

    ** Slide 2

    Text here
    - And bullets

    * Chapter Two