Updated the patch
Added support for current/total slide number field in the upper right of
each page (was bottom right, and previously it didn't work as it was off
by one).  There is now JavaScript code that dynamically counts the
slides and embeds the proper string.  This required that jQuery be split
out of the 'play.js' handler because 'slides.js' now uses it as well as
the new 'custom.js'.