Last set of tests
The next step will be to convert my existing slide deck to use the new
format.  Essentially, all lines starting with "**" will convert to "*",
and all lines with "*" will convert to "**".  I also need to check every
title to see if styling should be applied to any of the words.